add LD r8 r8 and HALT with the HALT BUG!

+++ src/tests.rs
@@ -1,7 +1,8 @@
-use std::fs::{create_dir, remove_dir_all};
+use std::fs::{create_dir, exists, remove_dir_all};
 use std::path::PathBuf;
 
 mod test_cpu_registers;
+mod test_halt;
 mod test_instructions;
 mod test_interrupts;
 mod test_mbc;
@@ -10,6 +11,8 @@ mod test_timer;
 
 pub fn setup_test_dir() {
     let test_dir = PathBuf::from("./test");
-    remove_dir_all(&test_dir).unwrap();
+    if exists(&test_dir).unwrap() {
+        remove_dir_all(&test_dir).unwrap();
+    }
     create_dir(&test_dir).unwrap();
 }
